requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
alembic==1.13.2
altair==5.4.1
annotated-types==0.7.0
ansicon==1.89.0
anyio==4.4.0
appdirs==1.4.4
asgiref==3.8.1
astor==0.8.1
asttokens==2.4.1
attrs==24.2.0
auth0-python==4.7.2
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
blessed==1.20.0
blinker==1.8.2
boto3==1.35.19
botocore==1.35.19
build==1.2.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.24
click==8.1.7
cohere==5.9.2
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.0
crewai==0.55.2
crewai-tools==0.12.0
cryptography==43.0.1
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.5
decorator==5.1.1
Deprecated==1.2.14
deprecation==2.1.0
distro==1.9.0
docker==7.1.0
docstring_parser==0.16
docx==0.2.4
docx2txt==0.8
editor==1.6.6
embedchain==0.1.121
executing==2.1.0
fastapi==0.114.2
fastavro==1.9.7
filelock==3.16.0
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.9.0
git-python==1.0.3
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.6
google-api-core==2.19.2
google-api-python-client==2.146.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.66.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.5
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-generativeai==0.7.2
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
gptcache==0.1.44
graphviz==0.20.3
greenlet==3.1.0
grpc-google-iam-v1==0.13.1
grpcio==1.66.1
grpcio-status==1.62.3
grpcio-tools==1.62.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
html2image==2.0.5
html2text==2024.2.26
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.24.7
humanfriendly==10.0
hyperframe==6.0.1
idna==3.9
importlib_metadata==8.4.0
importlib_resources==6.4.5
iniconfig==2.0.0
inquirer==3.4.0
instructor==1.3.3
ipykernel==6.29.5
ipython==8.27.0
jedi==0.19.1
Jinja2==3.1.4
jinxed==1.3.0
jiter==0.4.2
jmespath==1.0.1
joblib==1.4.2
jq==1.8.0
json_repair==0.25.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.7
kubernetes==30.1.0
lancedb==0.5.7
langchain==0.2.16
langchain-chroma==0.1.4
langchain-cohere==0.1.9
langchain-community==0.2.17
langchain-core==0.2.40
langchain-experimental==0.0.65
langchain-openai==0.1.25
langchain-text-splitters==0.2.4
langsmith==0.1.120
lark==1.2.2
litellm==1.46.4
lolviz==1.4.4
lxml==5.3.0
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mem0ai==0.0.20
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.8.1
nest-asyncio==1.6.0
nltk==3.9.1
nodeenv==1.9.1
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.19.2
open-interpreter==0.3.13
openai==1.45.0
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-exporter-otlp-proto-http==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
orjson==3.10.7
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.1
pandas==2.2.2
parameterized==0.9.0
parso==0.8.4
pillow==10.4.0
platformdirs==4.3.3
pluggy==1.5.0
portalocker==2.10.1
posthog==3.6.5
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==4.25.4
psutil==5.9.8
pulsar-client==3.5.0
pure_eval==0.2.3
py==1.11.0
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.1
pydantic-settings==2.5.2
pydantic_core==2.23.3
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.9.0
pylance==0.9.18
pyparsing==3.1.4
pypdf==4.3.1
pyperclip==1.9.0
PyPika==0.48.9
pyproject_hooks==1.1.0
pyreadline3==3.5.2
pyright==1.1.380
pysbd==0.3.4
PySocks==1.7.1
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytube==15.0.0
pytz==2024.2
pywin32==306
PyYAML==6.0.2
pyzmq==26.2.0
qdrant-client==1.11.1
ratelimiter==1.2.0.post0
readchar==4.2.0
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
retry==0.9.2
rich==13.8.1
rpds-py==0.20.0
rsa==4.9
runs==1.2.2
s3transfer==0.10.2
schema==0.7.7
selenium==4.24.0
semver==3.0.2
Send2Trash==1.8.3
setuptools==74.1.2
shapely==2.0.6
shellingham==1.5.4
shortuuid==1.0.13
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.34
stack-data==0.6.3
starlette==0.37.2
streamlit==1.38.0
sympy==1.13.2
tabulate==0.9.0
tenacity==8.5.0
termcolor==2.3.0
tiktoken==0.7.0
tokenizers==0.20.0
tokentrim==0.1.13
toml==0.10.2
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
trio==0.26.2
trio-websocket==0.11.1
typer==0.12.5
types-requests==2.32.0.20240914
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.30.6
watchdog==4.0.2
watchfiles==0.24.0
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==13.0.1
wget==3.2
wrapt==1.16.0
wsproto==1.2.0
xmod==1.8.1
yarl==1.11.1
yaspin==3.0.2
zipp==3.20.2
